设置消息唯一ID

+++ scripts/mysql/sqls/20220422190008_create_table_messages.sql
@@ -31,7 +31,7 @@ CREATE TABLE `messages` (
   `sender_avatar_url` varchar(255) CHARACTER SET utf8 DEFAULT '',
   `session_id` char(40) CHARACTER SET utf8 DEFAULT '' COMMENT '单例:会话ID,群聊:群ID',
   `session_type` tinyint DEFAULT '0' COMMENT '1:单聊 2:群聊',
-  `seq` bigint DEFAULT '0' COMMENT '会话消息唯一ID',
+  `seq` int DEFAULT '0' COMMENT '会话消息唯一ID',
   `msg_from` int DEFAULT '0' COMMENT '100:用户消息 200:系统消息',
   `content_type` int DEFAULT '0',
   `content` varchar(3000) CHARACTER SET utf8 DEFAULT '',
